pyproject.toml
[tool.poetry]
name = "blog-project"
version = "0.1.0"
description = ""
authors = ["Walex <w.alex36@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.36"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
flask-jwt-extended = "^4.7.1"
psycopg2-binary = "^2.9.10"
gunicorn = "^23.0.0"
flask-bcrypt = "^1.0.1"
flask-marshmallow = "^1.2.1"
marshmallow-sqlalchemy = "^1.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-mock = "^3.14.0"

[tool.pytest.ini_options]
addopts = "-ra -vvv"
testpaths = ["tests"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
